% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/lmmv.R
\name{lmmv}
\alias{lmmv}
\title{lmmv}
\usage{
lmmv(x, y)
}
\arguments{
\item{x}{numeric matrix or numeric data.frame. The first column must be 1, the constant.}

\item{y}{response for linear regression}
}
\value{
A list contains estimates of coefficient and variances of estimates.
}
\description{
This function uses Available Cases method to conduct linear regression for numeric data with missing value.
}
\examples{
n = 1000
x1 <- runif(n)
x2 <- runif(n)
x3 <- runif(n)
y <- runif(n)
idx1 = sample(n, round(0.1 * n), replace = FALSE)
idx2 = sample(n, round(0.1 * n), replace = FALSE)
idx3 = sample(n, round(0.1 * n), replace = FALSE)
idy = sample(n, round(0.1 * n), replace = FALSE)
x1[idx1] = NA
x2[idx2] = NA
x3[idx3] = NA
y[idy] = NA
# The first column of the Data matrix is 1.
#By Now, we only provide the model includes the intersect.
x = cbind(1,x1,x2,x3)
lmmv(x,y)
}

