% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/S3method.R
\name{plot.ACPCA}
\alias{plot.ACPCA}
\title{screen plot BML object}
\usage{
\method{plot}{ACPCA}(ACPCA, npc = "maximum")
}
\arguments{
\item{ACPCA}{ACpac object}

\item{npc}{Number of principle components to display.}
}
\description{
screen plot BML object
}

